+++ api/operator/v1beta1/vmagent_types.go
@@ -48,6 +48,9 @@ type VMAgentSpec struct {
 	// PodMetadata configures Labels and Annotations which are propagated to the vmagent pods.
 	// +optional
 	PodMetadata *EmbeddedObjectMetadata `json:"podMetadata,omitempty"`
+	// ManagedMetadata defines metadata that will be added to the all objects
+	// created by operator for the given CustomResource
+	ManagedMetadata *ManagedObjectsMetadata `json:"managedMetadata,omitempty"`
 	// LogLevel for VMAgent to be configured with.
 	// INFO, WARN, ERROR, FATAL, PANIC
 	// +optional
@@ -303,10 +306,6 @@ type VMAgentSpec struct {
 	CommonApplicationDeploymentParams `json:",inline,omitempty"`
 }
 
-func (cr *VMAgent) setLastMetadata(lam LastAppliedMetadata) {
-	cr.ParsedLastAppliedMetadata = lam
-}
-
 func (cr *VMAgent) setLastSpec(prevSpec VMAgentSpec) {
 	cr.ParsedLastAppliedSpec = &prevSpec
 }
@@ -476,8 +475,6 @@ type VMAgent struct {
 	Spec VMAgentSpec `json:"spec,omitempty"`
 	// ParsedLastAppliedSpec contains last-applied configuration spec
 	ParsedLastAppliedSpec *VMAgentSpec `json:"-" yaml:"-"`
-	//ParsedLastAppliedMetadata contains last-applied cr.metadata fields
-	ParsedLastAppliedMetadata LastAppliedMetadata `json:"-" yaml:"-"`
 
 	Status VMAgentStatus `json:"status,omitempty"`
 }
@@ -503,7 +500,7 @@ func (cr *VMAgent) AsOwner() []metav1.OwnerReference {
 	}
 }
 
-func (cr VMAgent) PodAnnotations() map[string]string {
+func (cr *VMAgent) PodAnnotations() map[string]string {
 	annotations := map[string]string{}
 	if cr.Spec.PodMetadata != nil {
 		for annotation, value := range cr.Spec.PodMetadata.Annotations {
@@ -513,11 +510,22 @@ func (cr VMAgent) PodAnnotations() map[string]string {
 	return annotations
 }
 
-func (cr VMAgent) AnnotationsFiltered() map[string]string {
-	return filterMapKeysByPrefixes(cr.ObjectMeta.Annotations, annotationFilterPrefixes)
+func (cr *VMAgent) AnnotationsFiltered() map[string]string {
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
+	dst := filterMapKeysByPrefixes(cr.ObjectMeta.Annotations, annotationFilterPrefixes)
+	if cr.Spec.ManagedMetadata != nil {
+		if dst == nil {
+			dst = make(map[string]string)
+		}
+		for k, v := range cr.Spec.ManagedMetadata.Annotations {
+			dst[k] = v
+		}
+	}
+
+	return dst
 }
 
-func (cr VMAgent) SelectorLabels() map[string]string {
+func (cr *VMAgent) SelectorLabels() map[string]string {
 	return map[string]string{
 		"app.kubernetes.io/name":      "vmagent",
 		"app.kubernetes.io/instance":  cr.Name,
@@ -526,7 +534,7 @@ func (cr VMAgent) SelectorLabels() map[string]string {
 	}
 }
 
-func (cr VMAgent) PodLabels() map[string]string {
+func (cr *VMAgent) PodLabels() map[string]string {
 	lbls := cr.SelectorLabels()
 	if cr.Spec.PodMetadata == nil {
 		return lbls
@@ -535,14 +543,21 @@ func (cr VMAgent) PodLabels() map[string]string {
 	return labels.Merge(cr.Spec.PodMetadata.Labels, lbls)
 }
 
-func (cr VMAgent) AllLabels() map[string]string {
+func (cr *VMAgent) AllLabels() map[string]string {
 	selectorLabels := cr.SelectorLabels()
 	// fast path
-	if cr.ObjectMeta.Labels == nil {
+	if cr.ObjectMeta.Labels == nil && cr.Spec.ManagedMetadata == nil {
 		return selectorLabels
 	}
-	crLabels := filterMapKeysByPrefixes(cr.ObjectMeta.Labels, labelFilterPrefixes)
-	return labels.Merge(crLabels, selectorLabels)
+	var result map[string]string
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
+	if cr.ObjectMeta.Labels != nil {
+		result = filterMapKeysByPrefixes(cr.ObjectMeta.Labels, labelFilterPrefixes)
+	}
+	if cr.Spec.ManagedMetadata != nil {
+		result = labels.Merge(result, cr.Spec.ManagedMetadata.Labels)
+	}
+	return labels.Merge(result, selectorLabels)
 }
 
 func (cr VMAgent) PrefixedName() string {