+++ api/operator/v1beta1/vmalertmanager_types.go
@@ -36,8 +36,6 @@ type VMAlertmanager struct {
 	Spec VMAlertmanagerSpec `json:"spec"`
 	// ParsedLastAppliedSpec contains last-applied configuration spec
 	ParsedLastAppliedSpec *VMAlertmanagerSpec `json:"-" yaml:"-"`
-	//ParsedLastAppliedMetadata contains last-applied cr.metadata fields
-	ParsedLastAppliedMetadata LastAppliedMetadata `json:"-" yaml:"-"`
 
 	// Most recent observed status of the VMAlertmanager cluster.
 	// Operator API itself. More info:
@@ -55,6 +53,9 @@ type VMAlertmanagerSpec struct {
 	// PodMetadata configures Labels and Annotations which are propagated to the alertmanager pods.
 	// +optional
 	PodMetadata *EmbeddedObjectMetadata `json:"podMetadata,omitempty"`
+	// ManagedMetadata defines metadata that will be added to the all objects
+	// created by operator for the given CustomResource
+	ManagedMetadata *ManagedObjectsMetadata `json:"managedMetadata,omitempty"`
 
 	// Templates is a list of ConfigMap key references for ConfigMaps in the same namespace as the VMAlertmanager
 	// object, which shall be mounted into the VMAlertmanager Pods.
@@ -199,10 +200,6 @@ type VMAlertmanagerSpec struct {
 	CommonApplicationDeploymentParams `json:",inline,omitempty"`
 }
 
-func (cr *VMAlertmanager) setLastMetadata(lam LastAppliedMetadata) {
-	cr.ParsedLastAppliedMetadata = lam
-}
-
 func (cr *VMAlertmanager) setLastSpec(prevSpec VMAlertmanagerSpec) {
 	cr.ParsedLastAppliedSpec = &prevSpec
 }
@@ -266,7 +263,7 @@ func (cr *VMAlertmanager) AsOwner() []metav1.OwnerReference {
 	}
 }
 
-func (cr VMAlertmanager) PodAnnotations() map[string]string {
+func (cr *VMAlertmanager) PodAnnotations() map[string]string {
 	annotations := map[string]string{}
 	if cr.Spec.PodMetadata != nil {
 		for annotation, value := range cr.Spec.PodMetadata.Annotations {
@@ -276,11 +273,21 @@ func (cr VMAlertmanager) PodAnnotations() map[string]string {
 	return annotations
 }
 
-func (cr VMAlertmanager) AnnotationsFiltered() map[string]string {
-	return filterMapKeysByPrefixes(cr.ObjectMeta.Annotations, annotationFilterPrefixes)
+func (cr *VMAlertmanager) AnnotationsFiltered() map[string]string {
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
+	dst := filterMapKeysByPrefixes(cr.ObjectMeta.Annotations, annotationFilterPrefixes)
+	if cr.Spec.ManagedMetadata != nil {
+		if dst == nil {
+			dst = make(map[string]string)
+		}
+		for k, v := range cr.Spec.ManagedMetadata.Annotations {
+			dst[k] = v
+		}
+	}
+	return dst
 }
 
-func (cr VMAlertmanager) SelectorLabels() map[string]string {
+func (cr *VMAlertmanager) SelectorLabels() map[string]string {
 	return map[string]string{
 		"app.kubernetes.io/name":      "vmalertmanager",
 		"app.kubernetes.io/instance":  cr.Name,
@@ -289,7 +296,7 @@ func (cr VMAlertmanager) SelectorLabels() map[string]string {
 	}
 }
 
-func (cr VMAlertmanager) PodLabels() map[string]string {
+func (cr *VMAlertmanager) PodLabels() map[string]string {
 	lbls := cr.SelectorLabels()
 	if cr.Spec.PodMetadata == nil {
 		return lbls
@@ -297,14 +304,21 @@ func (cr VMAlertmanager) PodLabels() map[string]string {
 	return labels.Merge(cr.Spec.PodMetadata.Labels, lbls)
 }
 
-func (cr VMAlertmanager) AllLabels() map[string]string {
+func (cr *VMAlertmanager) AllLabels() map[string]string {
 	selectorLabels := cr.SelectorLabels()
 	// fast path
-	if cr.ObjectMeta.Labels == nil {
+	if cr.ObjectMeta.Labels == nil && cr.Spec.ManagedMetadata == nil {
 		return selectorLabels
 	}
-	crLabels := filterMapKeysByPrefixes(cr.ObjectMeta.Labels, labelFilterPrefixes)
-	return labels.Merge(crLabels, selectorLabels)
+	var result map[string]string
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
+	if cr.ObjectMeta.Labels != nil {
+		result = filterMapKeysByPrefixes(cr.ObjectMeta.Labels, labelFilterPrefixes)
+	}
+	if cr.Spec.ManagedMetadata != nil {
+		result = labels.Merge(result, cr.Spec.ManagedMetadata.Labels)
+	}
+	return labels.Merge(result, selectorLabels)
 }
 
 // ConfigSecretName returns configuration secret name for alertmanager