+++ api/operator/v1beta1/vmextra_types.go
@@ -50,9 +50,8 @@ const (
 	SkipValidationValue      = "true"
 	AdditionalServiceLabel   = "operator.victoriametrics.com/additional-service"
 	// PVCExpandableLabel controls checks for storageClass
-	PVCExpandableLabel                = "operator.victoriametrics.com/pvc-allow-volume-expansion"
-	lastAppliedSpecAnnotationName     = "operator.victoriametrics/last-applied-spec"
-	lastAppliedMetadataAnnotationName = "operator.victoriametrics/last-applied-metadata"
+	PVCExpandableLabel            = "operator.victoriametrics.com/pvc-allow-volume-expansion"
+	lastAppliedSpecAnnotationName = "operator.victoriametrics/last-applied-spec"
 )
 
 const (
@@ -67,18 +66,22 @@ const (
 var SchemeGroupVersion = schema.GroupVersion{Group: "operator.victoriametrics.com", Version: "v1beta1"}
 
 var (
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
 	labelFilterPrefixes []string
 	// default ignored annotations
+	// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
 	annotationFilterPrefixes = []string{"kubectl.kubernetes.io/", "operator.victoriametrics.com/", "operator.victoriametrics/"}
 )
 
 // SetLabelAndAnnotationPrefixes configures global filtering for child labels and annotations
 // cannot be used concurrently and should be called only once at lib init
+// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
 func SetLabelAndAnnotationPrefixes(labelPrefixes, annotationPrefixes []string) {
 	labelFilterPrefixes = labelPrefixes
 	annotationFilterPrefixes = append(annotationFilterPrefixes, annotationPrefixes...)
 }
 
+// TODO: @f41gh7 deprecated at will be removed at v0.52.0 release
 func filterMapKeysByPrefixes(src map[string]string, prefixes []string) map[string]string {
 	dst := make(map[string]string, len(src))
 OUTER:
@@ -1033,19 +1036,9 @@ type ScrapeObjectStatus struct {
 	CurrentSyncError string `json:"-"`
 }
 
-// LastAppliedMetadata defines well-known object metadata fields
-// since last reconcile apply loop
-// it's needed to properly track labels and annotations changes
-// for child objects
-type LastAppliedMetadata struct {
-	Annotations map[string]string `json:"annotations,omitempty"`
-	Labels      map[string]string `json:"labels,omitempty"`
-}
-
 type objectWithLastAppliedState[T, ST any] interface {
 	GetAnnotations() map[string]string
 	setLastSpec(ST)
-	setLastMetadata(LastAppliedMetadata)
 }
 
 func parseLastAppliedState[T objectWithLastAppliedState[T, ST], ST any](cr T) error {
@@ -1058,34 +1051,16 @@ func parseLastAppliedState[T objectWithLastAppliedState[T, ST], ST any](cr T) er
 		return fmt.Errorf("cannot parse last applied spec annotation=%q, remove this annotation manually from object : %w", lastAppliedSpecAnnotationName, err)
 	}
 	cr.setLastSpec(dst)
-	var lam LastAppliedMetadata
-	lastAppliedJSON := cr.GetAnnotations()[lastAppliedMetadataAnnotationName]
-	if len(lastAppliedJSON) == 0 {
-		return nil
-	}
-	if err := json.Unmarshal([]byte(lastAppliedJSON), &lam); err != nil {
-		return fmt.Errorf("cannot parse last applied metadata annotation=%q, remove this annotation manually from object : %w", lastAppliedMetadataAnnotationName, err)
-	}
-	cr.setLastMetadata(lam)
 	return nil
 }
 
 // HasSpecChanges compares single spec with last applied single spec stored in annotation
 func hasStateChanges(crMeta metav1.ObjectMeta, spec any) (bool, error) {
-	labels, annotations := crMeta.GetLabels(), crMeta.GetAnnotations()
-	lastAppliedSpecJSON := annotations[lastAppliedSpecAnnotationName]
-	lastAppliedMetaJSON := annotations[lastAppliedMetadataAnnotationName]
-	if len(lastAppliedSpecJSON) == 0 || len(lastAppliedMetaJSON) == 0 {
+	lastAppliedSpecJSON := crMeta.GetAnnotations()[lastAppliedSpecAnnotationName]
+	if len(lastAppliedSpecJSON) == 0 {
 		return true, nil
 	}
-	mt := LastAppliedMetadata{Labels: filterMapKeysByPrefixes(labels, labelFilterPrefixes), Annotations: filterMapKeysByPrefixes(annotations, annotationFilterPrefixes)}
-	md, err := json.Marshal(mt)
-	if err != nil {
-		return false, err
-	}
-	if !bytes.Equal(md, []byte(lastAppliedMetaJSON)) {
-		return true, nil
-	}
+
 	instanceSpecData, err := json.Marshal(spec)
 	if err != nil {
 		return false, err
@@ -1102,12 +1077,7 @@ func lastAppliedChangesAsPatch(crMeta metav1.ObjectMeta, spec any) (client.Patch
 	if err != nil {
 		return nil, fmt.Errorf("possible bug, cannot serialize single specification as json :%w", err)
 	}
-	lam := LastAppliedMetadata{
-		Labels:      filterMapKeysByPrefixes(crMeta.Labels, labelFilterPrefixes),
-		Annotations: filterMapKeysByPrefixes(crMeta.Annotations, annotationFilterPrefixes),
-	}
-	dlam, err := json.Marshal(lam)
-	patch := fmt.Sprintf(`{"metadata":{"annotations":{%q: %q, %q: %q}}}`, lastAppliedSpecAnnotationName, data, lastAppliedMetadataAnnotationName, dlam)
+	patch := fmt.Sprintf(`{"metadata":{"annotations":{%q: %q }}}`, lastAppliedSpecAnnotationName, data)
 	return client.RawPatch(types.MergePatchType, []byte(patch)), nil
 
 }
@@ -1432,4 +1402,18 @@ type StatusMetadata struct {
 	// ObservedGeneration defines current generation picked by operator for the
 	// reconcile
 	ObservedGeneration int64 `json:"observedGeneration,omitempty"`
+}
+
+// ManagedObjectsMetadata contains Labels and Annotations
+type ManagedObjectsMetadata struct {
+	// Labels Map of string keys and values that can be used to organize and categorize
+	// (scope and select) objects.
+	// More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels
+	Labels map[string]string `json:"labels,omitempty"`
+
+	// Annotations is an unstructured key value map stored with a resource that may be
+	// set by external tools to store and retrieve arbitrary metadata. They are not
+	// queryable and should be preserved when modifying objects.
+	// More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations
+	Annotations map[string]string `json:"annotations,omitempty"`
 }